# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: API Instituto
  version: 1.0.0
  description: |
    API REST del Instituto.

    ### Forma de las respuestas

    **En exito el body es el recurso pelado**, sin ningun sobre: un objeto en
    las respuestas de un recurso y un array en los listados. Un listado vacio
    devuelve `200 []`, nunca un 404. Un `DELETE` devuelve `204` sin body.

    **En error** el body es un objeto `{ success: false, message, details? }`,
    igual para todos los endpoints porque sale del manejador centralizado
    (`src/middlewares/error.middleware.js`).

    Son formas distintas a proposito: para saber si algo salio bien mirá el
    **status HTTP**, no una propiedad del body.

    Las respuestas de error estan definidas una sola vez en
    `components/responses`: no las redefinas, usalas con `$ref`.
servers:
  - url: https://api-instituto.onrender.com/api
    description: Produccion
  - url: http://localhost:3000/api
    description: Desarrollo local

# Un tag por recurso. Agrupa los endpoints en la UI y define el orden en que
# aparecen.
...
paths:
...
  /carreras/{id}:
    get:
      tags: [Carreras]
      summary: Obtiene una carrera por id
      description: Devuelve el detalle de una carrera activa. Una carrera dada de baja devuelve 404, igual que si no existiera. El endpoint es publico.
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la carrera
          schema: { type: string }
          example: clc1a2b3c4d5e6f7g8h9i0j1
      responses:
        "200":
          description: La carrera pedida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Carrera" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    patch:
      tags: [Carreras]
      summary: Modifica una carrera
      description: |
        **Requiere token con rol ADMIN.**

        Actualización parcial: se manda solo lo que cambia.

        - omitir un campo → queda como estaba
        - mandarlo en `null` → se borra (solo los opcionales)
        - mandarlo en `""` → 400; para borrar se usa `null`

        **El `slug` no se puede editar** y tampoco cambia solo al cambiar el
        nombre, para no romper URLs que ya lo referencien.

        Si viene `sedes`, **reemplaza por completo** el conjunto de sedes
        asignadas: no agrega ni quita de a una. Mandar `[]` la deja sin
        ninguna.

        Una carrera **dada de baja devuelve 404** y no se puede modificar. Hoy
        no hay endpoint para reactivarla.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la carrera
          schema: { type: string }
          example: clc1a2b3c4d5e6f7g8h9i0j1
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CarreraUpdateInput" }
      responses:
        "200":
          description: La carrera ya modificada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Carrera" }
        "400":
          description: >-
            El body es inválido, vino vacío, o alguna de las sedes indicadas no
            existe.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: La carrera no existe o está dada de baja.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      tags: [Carreras]
      summary: Da de baja una carrera (baja lógica)
      description: |
        **Requiere token con rol ADMIN.**

        No borra nada: la fila queda con `activa: false`. A partir de acá la
        carrera desaparece de `GET /carreras`, su detalle devuelve 404 y **no
        se puede preinscribir gente a ella** (`POST /preinscripciones` da 409).

        Los vínculos con sedes y las preinscripciones que ya existían se
        conservan.

        **Es irreversible desde la API**: hoy no hay endpoint para reactivar
        una carrera dada de baja, a diferencia de sedes y álbumes.

        Para borrarla de verdad está `DELETE /carreras/{id}/definitivo`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la carrera
          schema: { type: string }
          example: clc1a2b3c4d5e6f7g8h9i0j1
      responses:
        "204":
          description: Dada de baja. Sin body.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: La carrera no existe o ya estaba dada de baja.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /contacto/{id}:
    get:
      tags: [Contacto]
      summary: Obtiene un mensaje de contacto por id
      description: |
        **Requiere token con rol ADMIN.**
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid del mensaje de contacto
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      responses:
        "200":
          description: El mensaje de contacto pedido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MensajeContacto" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    patch:
      tags: [Contacto]
      summary: Marca como leido o responde un mensaje de contacto
      description: |
        Actualizacion **parcial**: se manda solo lo que cambia (`estado` y/o
        `respuesta`). Hay que enviar al menos un campo, un body vacio es 400.

        Si se envia `respuesta` y el mensaje **todavia no tenia una respuesta
        previa**, se fija automaticamente `respondidoAt = now()` y
        `estado = RESPONDIDO`, sin importar que otro `estado` se haya mandado
        en el mismo body.

        **Requiere token con rol ADMIN.**
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid del mensaje de contacto
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MensajeContactoUpdateInput" }
      responses:
        "200":
          description: El mensaje de contacto ya modificado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MensajeContacto" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      tags: [Contacto]
      summary: Elimina un mensaje de contacto
      description: |
        Eliminacion **fisica**: la fila se borra de la base (no hay soft
        delete en este modelo). Si no existe devuelve 404. Si se elimina
        correctamente devuelve 204 sin body.

        **Requiere token con rol ADMIN.**
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid del mensaje de contacto
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      responses:
        "204":
          description: Eliminado. Sin body.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /albums/{id}:
    get:
      tags: [Albums]
      summary: Obtiene un álbum con todas sus fotos
      security:
        # Auth opcional: el endpoint es publico, pero si llega un token se usa
        # (lo necesita `incluirInactivos`). El `{}` es lo que lo marca como
        # opcional; sin la lista entera, Swagger UI NO manda el token aunque
        # hayas apretado Authorize, y el request sale anonimo.
        - {}
        - bearerAuth: []
      description: |
        Endpoint público. A diferencia del listado, este **sí** trae el array
        completo de `imagenes`, ordenado por fecha de subida.

        Un álbum dado de baja **devuelve 404**, igual que si no existiera: un
        403 estaría confirmando que existe. Con `incluirInactivos=true` se
        devuelve de todos modos, pero eso exige un token de ADMIN.
      parameters:
        - name: id
          in: path
          required: true
          description: cuid del álbum
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
        - name: incluirInactivos
          in: query
          required: false
          description: >-
            Con `true` devuelve el álbum aunque esté dado de baja. **Requiere
            token con rol ADMIN**: sin él, la respuesta es 403.
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: El álbum pedido, con sus imágenes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlbumDetalle" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    patch:
      tags: [Albums]
      summary: Modifica un álbum
      description: |
        **Requiere token con rol ADMIN.**

        Actualización parcial: se manda solo lo que cambia.

        - omitir un campo → queda como estaba
        - mandarlo en `null` → se borra (solo `descripcion` y `fecha`)
        - mandarlo en `""` → 400; para borrar se usa `null`

        **El `slug` no cambia aunque cambie el título**, a propósito: si
        cambiara, se romperían las URLs ya publicadas.

        Un álbum dado de baja devuelve 404: primero hay que reactivarlo.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid del álbum
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AlbumUpdateInput" }
      responses:
        "200":
          description: El álbum ya modificado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlbumDetalle" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      tags: [Albums]
      summary: Da de baja un álbum (baja lógica)
      description: |
        **Requiere token con rol ADMIN.**

        **Es también el switch de visibilidad pública.** No borra nada: la fila
        queda con `activo: false`, las fotos siguen en el servicio de imágenes
        y se recupera todo con `POST /albums/{id}/reactivar`.

        A partir de acá el álbum desaparece del listado público y su detalle
        devuelve 404, salvo para un ADMIN que pase `incluirInactivos=true`.

        Consecuencia a tener presente: **las fotos de un álbum dado de baja
        siguen ocupando espacio en la nube para siempre**. Para liberarlo hay
        que borrar las imágenes una por una con
        `DELETE /albums/{id}/imagenes/{imagenId}`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid del álbum
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      responses:
        "204":
          description: Dado de baja. Sin body.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: El álbum no existe o ya estaba dado de baja.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    # Es lo que valida `authenticate` en src/middlewares/auth.middleware.js.
    # Para probar endpoints protegidos desde la UI: boton "Authorize" y pegar
    # el token SIN el prefijo "Bearer " (la UI lo agrega sola).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Carrera:
      type: object
      description: Una carrera disponible en el instituto.
      required:
        - id
        - nombre
        - slug
        - descripcion
        - duracionAnios
        - tituloOtorgado
        - modalidad
        - activa
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: cuid generado por la base.
          examples: ["clc1a2b3c4d5e6f7g8h9i0j1"]

        nombre:
          type: string
          examples: ["Tecnicatura Superior en Programacion"]

        slug:
          type: string
          examples: ["tecnicatura-superior-en-programacion"]

        descripcion:
          type: [string, "null"]
          examples: ["Carrera orientada al desarrollo de software."]

        duracionAnios:
          type: [integer, "null"]
          examples: [3]

        tituloOtorgado:
          type: [string, "null"]
          examples: ["Tecnico Superior en Programacion"]

        modalidad:
          type: string
          enum: [PRESENCIAL, VIRTUAL, HIBRIDA]

        activa:
          type: boolean
          description: >-
            `false` significa que la carrera esta dada de baja. Por defecto
            las carreras dadas de baja no aparecen en los endpoints publicos.

        imagenUrl:
          type: ["string", "null"]
          format: uri
          description: >-
            Imagen representativa de la carrera, servida por CDN. `null` si
            todavia no se le cargo ninguna. Se gestiona con
            `PUT` y `DELETE /carreras/{id}/imagen`: **no se puede setear desde
            el POST ni el PATCH**. Para una miniatura, insertar la
            transformacion en la URL (`/upload/w_400,q_auto,f_auto/`).
          examples: ["https://res.cloudinary.com/demo/image/upload/v1/instituto/albums/carreras/abc.jpg"]

        createdAt:
          type: string
          format: date-time

        updatedAt:
          type: string
          format: date-time
...
    CarreraUpdateInput:
      type: object
      description: |
        Body del `PATCH /carreras/{id}`. Al menos un campo, y no acepta campos
        de más.

        - omitir → queda como estaba
        - `null` → borra el valor (solo los opcionales)
        - `""` → 400; para borrar se usa `null`

        `slug` no se puede modificar y tampoco cambia solo al cambiar el
        nombre.
      minProperties: 1
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 150
        descripcion:
          type: [string, "null"]
          minLength: 1
          maxLength: 2000
        duracionAnios:
          type: [integer, "null"]
          minimum: 1
          maximum: 15
        tituloOtorgado:
          type: [string, "null"]
          minLength: 1
          maxLength: 150
        modalidad:
          type: string
          enum: [PRESENCIAL, VIRTUAL, HIBRIDA]
        sedes:
          type: array
          description: >-
            **Reemplaza por completo** el conjunto de sedes asignadas, no
            agrega ni quita de a una. Mandar `[]` la deja sin ninguna.
          items: { type: string }
...
    Imagen:
      type: object
      description: |
        Una foto de un álbum. `url` es lo único que necesita el front; el
        identificador interno del servicio de imágenes no se expone, porque
        borrar es cosa de la API.
      required: [id, url, createdAt]
      properties:
        id: { type: string, example: clx9z8y7x6w5v4u3t2s1r0q9 }
        url:
          type: string
          format: uri
          description: URL pública de la imagen, servida por CDN.
          example: https://res.cloudinary.com/demo/image/upload/v1/instituto/albums/clx1a2/foto.jpg
        width:
          type: [integer, "null"]
          description: >-
            Ancho en píxeles. Sirve para que el front reserve el espacio antes
            de que cargue la foto y evite el salto de layout.
          example: 1600
        height:
          type: [integer, "null"]
          example: 1067
        formato:
          type: [string, "null"]
          example: jpg
        bytes:
          type: [integer, "null"]
          description: Peso del archivo original.
          example: 482913
        createdAt: { type: string, format: date-time }
    Album:
      type: object
      description: |
        Un álbum tal como sale en el **listado**: con su portada y el total de
        fotos, pero sin el array completo. Para las fotos, `AlbumDetalle`.
      required: [id, titulo, slug, activo, cantidadImagenes, createdAt, updatedAt]
      properties:
        id: { type: string, example: clx1a2b3c4d5e6f7g8h9i0j1 }
        titulo: { type: string, example: Acto de colación 2026 }
        slug:
          type: string
          description: >-
            Se genera del título y no se puede elegir ni modificar. Si el
            título se repite se agrega un sufijo (`-2`, `-3`, ...).
          example: acto-de-colacion-2026
        descripcion:
          type: [string, "null"]
          example: Fotos del acto de colación de la promoción 2026.
        fecha:
          type: [string, "null"]
          format: date-time
          description: Fecha del evento que retrata el álbum, no la de carga.
          example: "2026-08-15T18:00:00.000Z"
        activo:
          type: boolean
          description: >-
            Si es `false` el álbum está dado de baja: no aparece en el listado
            público y su detalle devuelve 404.
          example: true
        portadaId: { type: [string, "null"] }
        portada:
          description: La foto que se muestra en el listado. `null` si el álbum no tiene fotos.
          oneOf:
            - $ref: "#/components/schemas/Imagen"
            - type: "null"
        cantidadImagenes:
          type: integer
          description: Total de fotos del álbum.
          example: 24
        autorId: { type: [string, "null"] }
        autor:
          description: Quién lo creó. `null` si el usuario fue eliminado.
          oneOf:
            - type: object
              properties:
                id: { type: string }
                nombre: { type: string }
                apellido: { type: string }
            - type: "null"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AlbumDetalle:
      description: |
        Lo mismo que `Album` más el array completo de `imagenes`, ordenado por
        fecha de subida. Es lo que devuelve el detalle y todas las escrituras.
      allOf:
        - $ref: "#/components/schemas/Album"
        - type: object
          required: [imagenes]
          properties:
            imagenes:
              type: array
              items: { $ref: "#/components/schemas/Imagen" }
...
    AlbumUpdateInput:
      type: object
      description: |
        Body del PATCH. Al menos un campo, y no acepta campos de más.

        - omitir → queda como estaba
        - `null` → borra el valor
        - `""` → 400; para borrar se usa `null`

        `slug` no se puede modificar y tampoco cambia solo al cambiar el título.
      minProperties: 1
      properties:
        titulo:
          type: string
          minLength: 1
          maxLength: 200
        descripcion:
          type: [string, "null"]
          minLength: 1
          maxLength: 2000
        fecha:
          type: [string, "null"]
          format: date-time
...
    MensajeContacto:
      type: object
      description: Un mensaje enviado desde el formulario de contacto.
      required: [id, nombre, email, mensaje, estado, createdAt]
      properties:
        id:
          type: string
          description: cuid generado por la base.
          examples: ["clx1a2b3c4d5e6f7g8h9i0j1"]
        nombre:
          type: string
          examples: ["Maria Gomez"]
        email:
          type: string
          format: email
          examples: ["maria.gomez@example.com"]
        telefono:
          type: [string, "null"]
          examples: ["+54 351 4567890"]
        asunto:
          type: [string, "null"]
          examples: ["Consulta sobre inscripcion"]
        mensaje:
          type: string
          examples: ["Queria saber si todavia estan abiertas las inscripciones para el ciclo 2026."]
        estado:
          type: string
          enum: [NO_LEIDO, LEIDO, RESPONDIDO]
          description: >-
            Nace en `NO_LEIDO`. Pasa a `RESPONDIDO` automaticamente en cuanto
            se carga una `respuesta` por primera vez.
        respuesta:
          type: [string, "null"]
          description: Texto de la respuesta del admin. Null hasta que se responde.
          examples: ["Hola Maria, si, las inscripciones siguen abiertas hasta fin de mes."]
        respondidoAt:
          type: [string, "null"]
          format: date-time
          description: Se fija automaticamente al cargar la primera respuesta. Null hasta entonces.
        createdAt:
          type: string
          format: date-time
...
    MensajeContactoUpdateInput:
      type: object
      description: >-
        Body del PATCH /contacto/{id}. Todos los campos son opcionales, pero
        hay que mandar al menos uno. No acepta campos que no esten en esta
        lista.
      minProperties: 1
      additionalProperties: false
      properties:
        estado:
          type: string
          enum: [NO_LEIDO, LEIDO, RESPONDIDO]
          description: >-
            Se puede setear a mano, por ejemplo para marcar `LEIDO` sin
            responder todavia.
          examples: ["LEIDO"]
        respuesta:
          type: string
          minLength: 1
          maxLength: 2000
          description: >-
            Si el mensaje no tenia una respuesta previa, cargarla fuerza
            `estado = RESPONDIDO` y fija `respondidoAt`, aunque en el mismo
            body se haya mandado otro `estado`.
          examples: ["Hola Maria, si, las inscripciones siguen abiertas hasta fin de mes."]
...
    Error:
      type: object
      description: Forma de cualquier respuesta de error de la API.
      required: [success, message]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
          description: Descripcion del error, apta para mostrar al usuario.
          examples: ["El registro solicitado no existe"]
        details:
          description: Informacion extra. Solo aparece en algunos errores.
        stack:
          type: string
          description: >-
            Traza del error. SOLO se incluye cuando NODE_ENV=development;
            en produccion nunca se expone.
    ValidationIssue:
      type: object
      description: Un campo que no paso la validacion de Zod.
      required: [source, field, message]
      properties:
        source:
          type: string
          enum: [body, params, query]
          description: En que parte del request estaba el campo.
        field:
          type: string
          examples: ["email"]
        message:
          type: string
          examples: ["Invalid email address"]
    ValidationError:
      type: object
      description: >-
        Error 400 devuelto por el middleware `validate()`. A diferencia del
        resto, `details` trae la lista de campos que fallaron.
      required: [success, message, details]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
          examples: ["Datos de entrada invalidos"]
        details:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
  responses:
...
    BadRequest:
      description: Los datos enviados no pasaron la validacion.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Unauthorized:
      description: Falta el token, es invalido, expiro o la cuenta esta desactivada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Autenticado, pero el rol no alcanza para esta accion.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: El recurso no existe.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Error inesperado del servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"